Logout service

+++ myApp/src/pages/settings/settings.ts
@@ -4,9 +4,11 @@ import { NavController,App, ViewController } from 'ionic-angular';
 //entity
 import { Message } from '../../app/common/entity/message.entity';
 import { User } from '../../app/common/entity/user.entity';
+import { ToastController } from 'ionic-angular';
 //service
 import { MessageService } from '../../app/common/service/message.service';
 import { StorageService } from '../../app/common/service/storage.service';
+import { UserService } from '../../app/common/service/user.service';
 
 import { LoginPage } from '../login/login';
 
@@ -23,9 +25,11 @@ export class SettingsPage {
   constructor(
   	public navCtrl: NavController,
   	private messageService: MessageService,
-  	private storageService: StorageService,
+    private storageService: StorageService,
+  	private userService: UserService,
   	public viewCtrl: ViewController,
-    public appCtrl: App) {
+    public appCtrl: App,
+    public toastCtrl: ToastController) {
 
   }
     /**
@@ -37,8 +41,19 @@ export class SettingsPage {
   	}
 
   	logout(): void{
-  		this.storageService.remove("user");
-  		this.appCtrl.getRootNav().push(LoginPage);
+      this.userService.logout().then(data=>{
+        if(data.judge==0){
+          this.storageService.remove("user");
+          this.appCtrl.getRootNav().push(LoginPage);
+        }else{
+          let toast = this.toastCtrl.create({
+            message: '注销失败',
+            duration: 3000
+          });
+          toast.present();
+        }
+      })
+  		
   	}
   	ngOnInit(): void {
   		this.user = this.storageService.read<User>('user');
